Working on getting the dd variable working, update working
Seems like the variables are somehow linked and they use the shape
variable to make the size of memory that is returned (I am assuming that
they just return a pointer of a given size). Do not see initialized
value of the array for the dd var, need to run update() to get first
value.