Adding grids for new vars, running into error
Getting and index error. Thought it was a problem with the grid object
(using number 1,2 for new variables in the bmi files) but that doesn't
seem to be it. Likely need to go back and rip out code until problem
goes away. or go back a commit or two. Test case failing so code doesn't
finish compile.